fix random generator mocking
The value generated by a random generator is scaled down to match the
interval of the desired value. I assumed it was using modulo, not divide,
to scale down, but I was wrong. So I added a parameter to the mock to
specify the range of the desired values. It's a quirk to have a scale
of 1, and thus no scaling.